+++ android/src/main/java/com/inbrain/InBrainSurveysModule.java
@@ -1,8 +1,6 @@
 package com.inbrain;
 
 import android.app.Activity;
-import android.os.Handler;
-import android.os.Looper;
 
 import com.facebook.react.bridge.Arguments;
 import com.facebook.react.bridge.Promise;
@@ -34,7 +32,6 @@ import com.inbrain.sdk.model.Survey;
 import com.inbrain.sdk.model.SurveyCategory;
 import com.inbrain.sdk.model.SurveyFilter;
 import com.inbrain.sdk.model.InBrainSurveyReward;
-import android.util.Log;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -42,18 +39,15 @@ import java.util.List;
 
 import javax.annotation.Nullable;
 
+import static android.graphics.Color.parseColor;
+
 public class InBrainSurveysModule extends ReactContextBaseJavaModule implements InBrainCallback {
 
     private final ReactApplicationContext reactContext;
-    private HashMap sessionData;
-    private String sessionID;
 
     public InBrainSurveysModule(ReactApplicationContext reactContext) {
         super(reactContext);
         this.reactContext = reactContext;
-        //TMP
-        this.sessionID = null;
-        this.sessionData = null;
     }
 
     @Override
@@ -65,50 +59,19 @@ public class InBrainSurveysModule extends ReactContextBaseJavaModule implements 
     // ***** SET INBRAIN *****
     // ***********************
     @ReactMethod
-    public void setInBrain(final String apiClientId, final String clientSecret, final String userId, final Promise promise) {
-        try {
-            // Validate parameters
-            notNull("apiClientId", apiClientId);
-            notNull("clientSecret", clientSecret);
-            notNull("userId", userId);
-
-            // Set the listener
-            InBrain.getInstance().removeCallback(this);
-            InBrain.getInstance().addCallback(this);
-
-            // Needs to be in main thread
-            new Handler(Looper.getMainLooper()).post(new Runnable() {
-                @Override
-                public void run() {
-                    try {
-                        // Call Braintree sdk
-                        InBrain.getInstance().setInBrain(getReactApplicationContext(), apiClientId, clientSecret, true, userId);
-                        // Everything went well, resolve the promise
-                        promise.resolve(null);
-                    } catch (Exception e) {
-                        promise.reject("ERR_SET_INBRAIN", e.getMessage(), e);
-                    }
-                }
-            });
-
-        } catch (Exception e) {
-            promise.reject("ERR_SET_INBRAIN", e.getMessage(), e);
-        }
+    public void setInBrain(final String apiClientId, final String clientSecret) {
+        // Set the listener
+        InBrain.getInstance().removeCallback(this);
+        InBrain.getInstance().addCallback(this);
+        InBrain.getInstance().setInBrain(getReactApplicationContext(), apiClientId, clientSecret, true);
     }
 
     // **********************************
-    // ***** SET INBRAIN VALUES FOR *****
+    // ***** SET USER ID *****
     // **********************************
     @ReactMethod
-    public void setInBrainValuesFor(final String sessionId, final ReadableMap data, Promise promise) {
-
-        new InBrainSDKParamSetter<HashMap<String, String>>() {
-            @Override
-            public void setParam(HashMap<String, String> param) {
-                InBrain.getInstance().setInBrainValuesFor(sessionId, toHashMap(data));
-            }
-        }.apply(promise, "values", toHashMap(data), "ERR_SET_INBRAIN_VALUES");
-
+    public void setUserID( final String userId) {
+        InBrain.getInstance().setUserID(getReactApplicationContext(), userId);
     }
 
     // **********************************
@@ -116,10 +79,7 @@ public class InBrainSurveysModule extends ReactContextBaseJavaModule implements 
     // **********************************
     @ReactMethod
     public void setSessionID(final String sessionId) {
-        //tmp data until android sdk new version
-        this.sessionID = sessionId;
-        HashMap<String, String> data = this.sessionData;
-        InBrain.getInstance().setInBrainValuesFor(sessionId, data);
+        InBrain.getInstance().setSessionId(sessionId);
     }
 
     // **********************************
@@ -127,10 +87,7 @@ public class InBrainSurveysModule extends ReactContextBaseJavaModule implements 
     // **********************************
     @ReactMethod
     public void setDataOptions(final ReadableMap data) {
-        //tmp data until android sdk new version
-        this.sessionData = toHashMap(data);
-        String sessionId = this.sessionID;
-        InBrain.getInstance().setInBrainValuesFor(sessionId, toHashMap(data));
+        InBrain.getInstance().setDataOptions(toHashMap(data));
     }
 
     // ************************
@@ -300,7 +257,7 @@ public class InBrainSurveysModule extends ReactContextBaseJavaModule implements 
 
                         WritableArray categories = Arguments.createArray();
                         for (SurveyCategory category:survey.categories) {
-                           categories.pushInt(category.getId());
+                            categories.pushInt(category.getId());
                         }
 
                         map.putArray("categories", categories);
@@ -362,116 +319,50 @@ public class InBrainSurveysModule extends ReactContextBaseJavaModule implements 
         }
     }
 
-    // **************************
-    // ***** SET VIEW TITLE *****
-    // **************************
-    @ReactMethod
-    public void setTitle(final String title, Promise promise) {
-
-        new InBrainSDKParamSetter<String>() {
-            @Override
-            public void setParam(String param) {
-                // This method doesn't exist in the Android SDK as setting the title is part of setNavigationBarConfig
-                // This is just here so we don't have to add any condition in the RN code.
-            }
-        }.apply(promise, "title", title, "ERR_SET_TITLE");
-
-    }
-
-    // *************************************
+    // ************************************
     // ***** SET NAVIGATION BAR CONFIG *****
     // *************************************
     @ReactMethod
-    public void setNavigationBarConfig(final ReadableMap configMap, Promise promise) {
-        final ToolBarConfig config = new ToolBarConfig();
+    public void setNavigationBarConfig(@Nullable String backgroundHex, @Nullable String buttonsHex,
+                                       @Nullable String titleHex, @Nullable String title,
+                                       @Nullable boolean hasShadow)  {
+        ToolBarConfig config = new ToolBarConfig();
+        
+        if (backgroundHex != null) {
+            int backgroundColor = parseColor(backgroundHex);
+            config.setToolbarColor(backgroundColor);
+        }
 
+        if (buttonsHex != null) {
+            int buttonsColor = parseColor(buttonsHex);
+            config.setBackButtonColor(buttonsColor);
+        }
+        
+        if (titleHex != null) {
+            int titleColor = parseColor(titleHex);
+            config.setTitleColor(titleColor);
+        }
+            
+        config.setTitle(title);
+        config.setElevationEnabled(hasShadow);
 
-        // Forwarding to SDK
-        new InBrainSDKParamSetter<ToolBarConfig>() {
-            @Override
-            public void setParam(ToolBarConfig param) {
-
-                // Extract parameters
-                new ColorMapParamExtractor(configMap, "backgroundColor") {
-
-                    @Override
-                    public void setParam(Integer color) {
-                        config.setToolbarColor(color);
-                    }
-                };
-                new ColorMapParamExtractor(configMap, "buttonsColor") {
-
-                    @Override
-                    public void setParam(Integer color) {
-                        config.setBackButtonColor(color);
-                    }
-                };
-                new ColorMapParamExtractor(configMap, "titleColor") {
-
-                    @Override
-                    public void setParam(Integer color) {
-                        config.setTitleColor(color);
-                    }
-                };
-                new BooleanMapParamExtractor(configMap, "hasShadow") {
-
-                    @Override
-                    public void setParam(Boolean elevationEnabled) {
-                        config.setElevationEnabled(elevationEnabled);
-                    }
-                };
-                new StringMapParamExtractor(configMap, "title") {
-
-                    @Override
-                    public void setParam(String title) {
-                        config.setTitle(title);
-                    }
-                };
-
-                // Finally call InBrain
-                InBrain.getInstance().setToolbarConfig(param);
-            }
-        }.apply(promise, "navigationBarConfig", config, "ERR_SET_NAVIGATION_BAR_CONFIG");
-
+        InBrain.getInstance().setToolbarConfig(config);
     }
-
 
     // *********************************
     // ***** SET STATUS BAR CONFIG *****
     // *********************************
     @ReactMethod
-    public void setStatusBarConfig(final ReadableMap configMap, Promise promise) {
-        final StatusBarConfig config = new StatusBarConfig();
-
-
-        // Forwarding to SDK
-        new InBrainSDKParamSetter<StatusBarConfig>() {
-            @Override
-            public void setParam(StatusBarConfig param) {
-
-                // Extract parameters
-                new BooleanMapParamExtractor(configMap, "lightStatusBar") {
-
-                    @Override
-                    public void setParam(Boolean lightStatusBar) {
-                        config.setStatusBarIconsLight(lightStatusBar);
-                    }
-                };
-                new ColorMapParamExtractor(configMap, "statusBarColor") {
-
-                    @Override
-                    public void setParam(Integer color) {
-                        config.setStatusBarColor(color);
-                    }
-                };
-
-                // Finally call InBrain
-                InBrain.getInstance().setStatusBarConfig(param);
-            }
-        }.apply(promise, "statusBarConfig", config, "ERR_SET_STATUS_BAR_CONFIG");
-
+    public void setStatusBarConfig(final boolean lightStatusBar, @Nullable String backgroundHex) {
+        StatusBarConfig config = new StatusBarConfig();
+        config.setStatusBarIconsLight(lightStatusBar);
+        if (backgroundHex != null) {
+            int backgroundColor = parseColor(backgroundHex);
+            config.setStatusBarColor(backgroundColor);
+        }
+        
+        InBrain.getInstance().setStatusBarConfig(config);
     }
-
 
     @ReactMethod
     public void addListener(String eventName) {
@@ -488,14 +379,12 @@ public class InBrainSurveysModule extends ReactContextBaseJavaModule implements 
     // ***********************
     @ReactMethod
     public void setLanguage(final String language, Promise promise) {
-
         new InBrainSDKParamSetter<String>() {
             @Override
             public void setParam(String param) {
                 InBrain.getInstance().setLanguage(language);
             }
         }.apply(promise, "language", language, "ERR_SET_LANGUAGE");
-
     }
 
     // ********************