+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+##[2.9.0] 2023-02-07
+
+### Added
+- `setInBrain` the method for initial inBrain SDK configuration;
+- `setUserID` the method to set userID separetely from initialization of the sdk; 
+- `setSessionID` the method to set the value to track user session;
+- `setStatusBarConfig` the method to customize Status Bar;
+- `setNavigationBarConfig` the method to customize Navigation Bar;
+- `StatusBarConfig` type, represents Status Bar configuration;
+- `NavigationBarConfig` type, represents Navigation Bar configuration.
+
+### Changed
+- `init` method deprecated;
+- `setSessionParameters` method deprecated;
+- `ConfigOptions` type deprecated;
+- `NavitagionBarOptions` type deprecated;
+- `StatusBarOptions` type deprecated;
+- `StylingOptionType` type deprecated.
+
 ##[2.8.0] 2023-01-02
 
 ### Added